refactor usermessage
