+++ src/userMessage.js
@@ -1,32 +1,12 @@
-const { DF_KEY, GOOGLE_API_1, GOOGLE_API_2 } = require('../config');
+const { DF_KEY } = require('../config');
 const apiai = require('apiai');
-const request = require('request-promise');
 const saveConversation = require('./database/queries/save_conversation');
 const saveMessage = require('./database/queries/save_message');
-const gSheetLookup = require('./googleSheetRef');
+const googleCall = require('./googleCall');
 
 const app = apiai(DF_KEY);
 
-const getResource = (countryObj) => {
-  const cellRef = gSheetLookup[countryObj.lookup];
-  const url = GOOGLE_API_1 + cellRef + GOOGLE_API_2;
-
-  return request(url)
-    .then((body) => {
-      const resourceArray = JSON.parse(body).values.map((resource) => {
-        const singleResourceArray = resource.map((str, index, array) => {
-          if (index % 2 === 1) return null;
-          return { text: array[index], href: array[index + 1] };
-        });
-
-        return singleResourceArray.filter(Boolean);
-      });
-      return [].concat(...resourceArray);
-    });
-};
-
 const apiaiCall = (req, res, speech) => {
-  saveMessage(speech, req.body.uniqueId);
   const requestdf = app.textRequest(speech, {
     sessionId: req.body.uniqueId,
   });
@@ -55,11 +35,7 @@ const apiaiCall = (req, res, speech) => {
     if (payload.resources) {
       let { selectedCountries } = req.body;
       selectedCountries = selectedCountries || [{ lookup: 'Global' }];
-
-      const promiseArray = selectedCountries.map(async (countryObj) => {
-        const resource = await getResource(countryObj);
-        return resource;
-      });
+      const promiseArray = googleCall(selectedCountries);
 
       Promise.all(promiseArray)
         .then((resources2dArray) => {
@@ -95,9 +71,10 @@ const apiaiCall = (req, res, speech) => {
 };
 
 const userMessage = (req, res) => {
-  const { speech } = req.body;
+  const { speech, uniqueId } = req.body;
+  saveConversation(uniqueId);
+  saveMessage(speech, uniqueId);
   apiaiCall(req, res, speech);
-  saveConversation(req.body.uniqueId);
 };
 
 module.exports = userMessage;
